Fix potential key buffer overflow condition
A key of <24 bytes (which is not checked elsewhere) can cause "kp" be accessed past the end of key data, for data lengths not a multiple of 8. This change causes the key offset "off" rotation to repeat until "kp" is safely within the key data.

Note the same rotation at line 135 does not have the same issue because the "while" condition at line 140 has "&& (kp < kpend)" and will bypass doing any access until "off" is rotated again and "kp" is within the key data.